Choice Modeling
For each travel purpose, hybrid multinomial logit and hybrid nested logit models are estimated. It should be noted that the measurement equation models and the hybrid choice models are estimated sequentially. The results of the measurement. hybrid multinomial logit model serves as a basis, and the hybrid nested logit tests for the similarity among the existing alternatives in contrast to the new emerging mode. The PandasBiogeme software package (Bierlaire, 2023) was used for the estimation of the models.